Update embedding column tests
